Change primary key type to int where it was long

+++ backend/api/Models/ExceptionLog.cs
@@ -1,10 +1,8 @@
-using System.Net;
-
 namespace api.Models;
 
 public class ExceptionLog
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public required DateTime UtcTimestamp { get; set; }
     public required int HttpStatusCode { get; set; }
     public required string Method { get; set; }
